+++ examples/readme.py
@@ -1,8 +1,8 @@
 import numpy as np
 from sklearn.ensemble import RandomForestClassifier
 
+from ocean import MixedIntegerProgramExplainer
 from ocean.datasets import load_adult
-from ocean.explainers import MixedIntegerProgramExplainer
 
 # Load the adult dataset
 (data, target), mapper = load_adult()
